Separated development of bot
Interest has been expressed in having the bot go public and so I am separating the bot's development into 2 groups. The first is for a Discord known as the "Pumpki Patch" and will include commands specific to that Discord, and the other is for general public use with no commands dedicated to a specific Discord.